do not reset linked components on revert to draft

+++ services/apps/alcs/src/alcs/application-decision/application-decision-v2/application-decision/application-decision-v2.service.ts
@@ -412,7 +412,7 @@ export class ApplicationDecisionV2Service {
           existingDecision.components ?? [],
           false,
         );
-    } else if (existingDecision.conditions) {
+    } else if (updateDto.conditions === null && existingDecision.conditions) {
       await this.decisionConditionService.remove(existingDecision.conditions);
     }
   }
